monitoring/otel_proxy/requirements.txt
Flask>=3.1.0
PyYAML>=6.0.2
opentelemetry-api>=1.30.0
opentelemetry-proto>=1.30.0
protobuf>=5.29.3
protobuf>=5.29.3
requests>=2.32.3